requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
pycodestyle==2.8.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
